D&D/backend/requirements.txt
# 🏔️ Middle Earth RPG - Backend Requirements
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow==10.2.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
django-storages==1.14.2
boto3==1.34.34
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
